[Cleanup] [Multiton-API] [No Instance-Parameter only allowed if 1 instance]
- If there are more than 1 instances, you need to pass the instance number
  explicitly now. Only for one instance the call getXyzInstance() is allowed.

Reviewed by M. Wodok

+++ library/xgpl_src/IsoAgLib/comm/Part10_TaskController_Client/impl/tcclient_c.cpp
@@ -31,7 +31,7 @@
 namespace __IsoAgLib {
 
   TcClient_c&
-  getTcClientInstance( uint8_t instance ) {
+  getTcClientInstance( unsigned instance ) {
     MACRO_MULTITON_GET_INSTANCE_BODY( TcClient_c, PRT_INSTANCE_CNT, instance );
   }
